修改几个bug 1.judge改为Judge 2.ouput模式下PA8引脚与LED同一电平输出

+++ MDK-ARM/Users/KEY.c
@@ -41,7 +41,7 @@ typedef struct
 	uint8_t Time;
 	uint8_t Press;
 	uint8_t PressTime;
-	uint8_t Count;
+	int Count;
 } KEY_TypeDef;
 
 static KEY_TypeDef Keys[KEYn] = {0};
